# openapi.yml
openapi: 3.1.0
info:
  title: MemoryOS browser API
  description: The same-origin contract consumed by the MemoryOS browser application.
  version: 1.0.0
servers:
- url: /
  description: Same-origin MemoryOS runtime
paths:
...
  /api/users/{actorId}/groups:
    post:
      tags:
      - Users
      summary: Replace a user's ordinary Group memberships
      operationId: replaceUserGroups
      parameters:
      - name: X-MemoryOS-CSRF
        in: header
        description: Same-origin non-simple request guard for browser-session mutations.
        required: true
        schema:
          type: string
          enum:
          - "1"
      - name: actorId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReplaceUserGroupsRequest"
        required: true
      responses:
        "204":
          description: The user's ordinary Group memberships were replaced
        "401":
          description: No accepted authentication is present
        "403":
          description: The actor lacks SYSTEM_ADMIN authority or the same-origin header
            is missing
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblem"
        "404":
          description: The user or a requested Group does not exist in the current
            Tenant
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblem"
      security:
      - browserSession: []
      - bearerAuth: []
  /api/users/{actorId}/deactivate:
    post:
      tags:
      - Users
      summary: Deactivate an existing Tenant member
      operationId: deactivateUser
      parameters:
      - name: X-MemoryOS-CSRF
        in: header
        description: Same-origin non-simple request guard for browser-session mutations.
        required: true
        schema:
          type: string
          enum:
          - "1"
      - name: actorId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: The member is inactive
        "401":
          description: No accepted authentication is present
        "403":
          description: "The actor lacks USERS_MANAGE authority, the target is protected,\
            \ or the same-origin header is missing"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblem"
        "404":
          description: The member does not exist in the current Tenant
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblem"
      security:
      - browserSession: []
      - bearerAuth: []
  /api/users/{actorId}/activate:
    post:
      tags:
      - Users
      summary: Activate an existing Tenant member
      operationId: activateUser
      parameters:
      - name: X-MemoryOS-CSRF
        in: header
        description: Same-origin non-simple request guard for browser-session mutations.
        required: true
        schema:
          type: string
          enum:
          - "1"
      - name: actorId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: The member is active
        "401":
          description: No accepted authentication is present
        "403":
          description: "The actor lacks USERS_MANAGE authority, the target is protected,\
            \ or the same-origin header is missing"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblem"
        "404":
          description: The member does not exist in the current Tenant
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblem"
      security:
      - browserSession: []
      - bearerAuth: []
...
  /api/users:
    get:
      tags:
      - Users
      summary: List users manageable by the current IAM administrator
      operationId: listUsers
      parameters:
      - name: search
        in: query
        required: false
        schema:
          type: string
          maxLength: 200
          minLength: 0
      - name: status
        in: query
        required: false
        schema:
          type: string
          enum:
          - ACTIVE
          - INACTIVE
          - INVITED
      - name: role
        in: query
        required: false
        schema:
          type: string
          enum:
          - OWNER
          - MEMBER
      - name: groupId
        in: query
        required: false
        schema:
          type: string
          format: uuid
      - name: sort
        in: query
        required: false
        schema:
          type: string
          default: NAME_ASC
          enum:
          - NAME_ASC
          - NAME_DESC
          - EMAIL_ASC
          - EMAIL_DESC
          - STATUS_ASC
          - STATUS_DESC
          - ROLE_ASC
          - ROLE_DESC
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
          minimum: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 20
          maximum: 100
          minimum: 1
      responses:
        "200":
          description: A bounded page of current memberships and eligible pending
            invitations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPage"
        "400":
          description: Invalid user directory query
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblem"
        "401":
          description: No accepted authentication is present
        "403":
          description: The actor lacks USERS_MANAGE authority
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ApiProblem"
      security:
      - browserSession: []
      - bearerAuth: []
...
components:
  schemas:
...
    ReplaceUserGroupsRequest:
      type: object
      additionalProperties: false
      properties:
        groupIds:
          type: array
          items:
            type: string
            format: uuid
          maxItems: 100
          maxLength: 100
          minItems: 0
      required:
      - groupIds
...
    GroupSystemKey:
      type: string
      enum:
      - ADMIN
      - BASIC
...
    AccountType:
      type: string
      enum:
      - STANDARD
    TenantMembershipRole:
      type: string
      enum:
      - OWNER
      - MEMBER
    UserCounts:
      type: object
      additionalProperties: false
      properties:
        active:
          type: integer
          format: int64
          minimum: 0
        inactive:
          type: integer
          format: int64
          minimum: 0
        invited:
          type: integer
          format: int64
          minimum: 0
      required:
      - active
      - inactive
      - invited
    UserGroup:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        systemKey:
          oneOf:
          - $ref: "#/components/schemas/GroupSystemKey"
          - type: "null"
      required:
      - id
      - name
      - systemKey
    UserListItem:
      type: object
      additionalProperties: false
      properties:
        actorId:
          type:
          - string
          - "null"
          format: uuid
          description: Stable Actor identifier for a membership row; null for an invitation
            row.
        invitationId:
          type:
          - string
          - "null"
          format: uuid
          description: Invitation identifier for an invited row; null for a membership
            row.
        displayName:
          type:
          - string
          - "null"
        email:
          type:
          - string
          - "null"
        emailVerified:
          type:
          - boolean
          - "null"
        profileIssuer:
          type:
          - string
          - "null"
        role:
          oneOf:
          - $ref: "#/components/schemas/TenantMembershipRole"
          - type: "null"
        accountType:
          oneOf:
          - $ref: "#/components/schemas/AccountType"
          - type: "null"
        status:
          type: string
          enum:
          - ACTIVE
          - INACTIVE
          - INVITED
        groups:
          type: array
          items:
            $ref: "#/components/schemas/UserGroup"
        invitationExpiresAt:
          type:
          - string
          - "null"
          format: date-time
          description: Invitation expiry for an invited row; null for a membership
            row.
      required:
      - accountType
      - actorId
      - displayName
      - email
      - emailVerified
      - groups
      - invitationExpiresAt
      - invitationId
      - profileIssuer
      - role
      - status
    UserPage:
      type: object
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/UserListItem"
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          maximum: 100
          minimum: 1
        totalItems:
          type: integer
          format: int64
          minimum: 0
        totalPages:
          type: integer
          format: int64
          minimum: 0
        counts:
          $ref: "#/components/schemas/UserCounts"
      required:
      - counts
      - items
      - page
      - size
      - totalItems
      - totalPages
...
    ApiProblem:
      type: object
      additionalProperties: false
      properties:
        type:
          type: string
          format: uri
          description: Stable problem type for capability failures; omitted means
            RFC 9457 `about:blank`.
        title:
          type: string
          description: Short human-readable problem category.
        status:
          type: integer
          format: int32
          description: HTTP status code.
        detail:
          type: string
          description: Safe human-readable fallback detail.
        instance:
          type: string
          format: uri-reference
          description: Request path that produced the problem.
        code:
          type: string
          description: Stable capability-prefixed code; present only for expected
            capability failures.
          pattern: "^[A-Z][A-Z0-9_]*$"
        errors:
          type: array
          description: Field validation failures with safe allowlisted parameters;
            no rejected values.
          items:
            type: object
            additionalProperties: false
            properties:
              field:
                type: string
              message:
                type: string
                description: Compatible fallback message; clients localize by code.
              code:
                type: string
                enum:
                - REQUIRED
                - INVALID
                - EMAIL
                - SIZE
                - MIN
                - MAX
              params:
                type: object
                additionalProperties: false
                properties:
                  min:
                    type: number
                  max:
                    type: number
            required:
            - code
            - field
            - message
            - params
      required:
      - detail
      - instance
      - status
      - title
  securitySchemes:
    browserSession:
      type: apiKey
      description: HttpOnly JDBC-backed Spring Security session cookie.
      name: SESSION
      in: cookie
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT